+++ webviz_subsurface/plugins/_parameter_distribution.py
@@ -152,6 +152,8 @@ and the parameter columns.
                             ],
                             value=self.parameter_columns[0],
                             clearable=False,
+                            persistence=True,
+                            persistence_type="session",
                         ),
                         self.make_buttons(self.ids("prev-btn"), self.ids("next-btn")),
                     ],