+++ webviz_subsurface/plugins/_well_cross_section_fmu.py
@@ -180,6 +180,8 @@ e.g. [xtgeo](https://xtgeo.readthedocs.io/en/latest/).
                         ],
                         value=self.wellfiles[0],
                         clearable=False,
+                        persistence=True,
+                        persistence_type="session",
                     ),
                 ]
             )
@@ -200,6 +202,8 @@ e.g. [xtgeo](https://xtgeo.readthedocs.io/en/latest/).
                         value=self.surfacenames,
                         clearable=True,
                         multi=True,
+                        persistence=True,
+                        persistence_type="session",
                     ),
                 ]
             ),
@@ -221,6 +225,8 @@ e.g. [xtgeo](https://xtgeo.readthedocs.io/en/latest/).
                         value=list(self.ensembles.keys())[0],
                         clearable=False,
                         multi=False,
+                        persistence=True,
+                        persistence_type="session",
                     ),
                 ]
             ),
@@ -244,6 +250,8 @@ e.g. [xtgeo](https://xtgeo.readthedocs.io/en/latest/).
                             ],
                             value=self.segyfiles[0],
                             clearable=False,
+                            persistence=True,
+                            persistence_type="session",
                         ),
                     ]
                 ),
@@ -267,13 +275,15 @@ e.g. [xtgeo](https://xtgeo.readthedocs.io/en/latest/).
                             ],
                             placeholder="Display log",
                             clearable=True,
+                            persistence=True,
+                            persistence_type="session",
                         ),
                     ]
                 ),
             )
         return html.Div(
             style={"visibility": "hidden"},
-            children=dcc.Dropdown(id=self.ids("marginal-log")),
+            children=dcc.Dropdown(id=self.ids("marginal-log"), persistence=True),
         )
 
     @property
@@ -299,6 +309,8 @@ e.g. [xtgeo](https://xtgeo.readthedocs.io/en/latest/).
                     debounce=True,
                     type="number",
                     value=self.sampling,
+                    persistence=True,
+                    persistence_type="session",
                 ),
                 html.Label("Extension"),
                 dcc.Input(
@@ -306,8 +318,16 @@ e.g. [xtgeo](https://xtgeo.readthedocs.io/en/latest/).
                     debounce=True,
                     type="number",
                     value=self.nextend,
+                    persistence=True,
+                    persistence_type="session",
                 ),
-                dcc.Checklist(id=self.ids("options"), options=options, value=value),
+                dcc.Checklist(
+                    id=self.ids("options"),
+                    options=options,
+                    value=value,
+                    persistence=True,
+                    persistence_type="session",
+                ),
             ],
         )
 
@@ -323,6 +343,8 @@ e.g. [xtgeo](https://xtgeo.readthedocs.io/en/latest/).
                     value=self.surfacenames[0],
                     clearable=False,
                     multi=False,
+                    persistence=True,
+                    persistence_type="session",
                 ),
                 dcc.Dropdown(
                     id=self.ids("surface-type"),
@@ -340,6 +362,8 @@ e.g. [xtgeo](https://xtgeo.readthedocs.io/en/latest/).
                     value="stddev",
                     clearable=False,
                     multi=False,
+                    persistence=True,
+                    persistence_type="session",
                 ),
             ],
         )
@@ -394,7 +418,9 @@ e.g. [xtgeo](https://xtgeo.readthedocs.io/en/latest/).
                             id=self.ids("show_map"),
                             children="Show map",
                         ),
-                        dcc.Store(id=self.ids("fencespec"), data=[]),
+                        dcc.Store(
+                            id=self.ids("fencespec"), storage_type="session", data=[]
+                        ),
                     ],
                 ),
             ],